smart_steps_ai/requirements.txt
anthropic>=0.13.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
typer>=0.9.0
rich>=13.0.0
matplotlib>=3.7.0
PyJWT>=2.8.0
openai>=1.5.0
tiktoken>=0.5.0
